+++ Makefile
@@ -1,6 +1,6 @@
 NAME		= miniRT
 CC			= gcc -g
-# CFLAGS		= -Wall -Wextra -Werror
+CFLAGS		= -Wall -Wextra -Werror
 RM			= /bin/rm -f
 
 LIBFT_DIR	= ./libft
