+++ files/sphere.rt
@@ -7,7 +7,7 @@ L   0,-5,10                      1.0         0,255,0
 
 pl  0,1,0   0,-1,0               255,255,255
 
-sp  0,0,0                       1           255,255,0  +checkerboard m0.5,20
+sp  0,0,0                       1           255,255,0 m0.5,5
 # sp  0,0,0                       1           255,255,0
 # sp  -2,0,0                      1           255,0,255
 # sp  2,0,0                       1           0,255,255
