+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: sharnvon <sharnvon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/24 16:31:44 by sharnvon          #+#    #+#             */
-/*   Updated: 2023/03/24 16:54:02 by sharnvon         ###   ########.fr       */
+/*   Updated: 2023/03/26 04:08:09 by sharnvon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,10 +28,6 @@ int	main(int argc, char **argv)
 	initial(&data);
 	render_scene(&data);
 	window_interface(&data, NONE);
-	// Test texture
-	// data.scene.img = txtr_checker_test(&data);
-	// mlx_put_image_to_window(data.mlx, data.win, data.scene.img.ptr, 0, 0);
-
 	mlx_hook(data.win, X_EVENT_KEY_EXIT, 1L << 0, &rt_close, &data);
 	mlx_hook(data.win, 2, 1L << 0, &keyhandler_press, &data);
 	mlx_hook(data.win, 3, 1L << 1, &keyhandler_release, &data);
